Create api flow for accounts page
